requirements.txt
torch==2.4.0
torchvision>=0.19.0
torchaudio>=2.4.0

diffusers==0.31.0
transformers==4.44.2
accelerate==0.33.0
safetensors>=0.4.3


huggingface-hub==0.24.6
hf_transfer==0.1.7

fastapi==0.112.2
uvicorn
python-multipart==0.0.9
pydantic>=2.7,<3
pydantic_settings

Pillow>=10.3.0
opencv-python-headless>=4.9.0.80
numpy>=1.26.4
imageio>=2.34.0
imageio-ffmpeg>=0.4.9

protobuf>=4.24.0
sentencepiece>=0.2.0
omegaconf>=2.3.0
einops>=0.7.0
scipy>=1.10.0
ftfy>=6.2.0
regex>=2024.7.24
